Feat: fix code according to comments and fixed middleware test suits
