+++ src/index.js
@@ -1,7 +1,7 @@
 //create an express server
 const express = require('express');
-const router = require('./routes/user.routes');
-const authRouter = require('./routes/login.routes');
+const router = require('./routes/user.route');
+const authRouter = require('./routes/login.route');
 require('dotenv').config();
 const projectRouter = require('./routes/project.route');
