--- src/middlewares/login.validation.js
@@ -1,9 +1,0 @@
-const { loginReqSchema } = require('./schemas.validator');
-exports.validateLoginReq = (req, res, next) => {
-  const { error } = loginReqSchema.validate(req.body);
-  if (error) {
-    res.status(400).json({ message: error.message });
-  } else {
-    next();
-  }
-};
