+++ src/routes/login.route.js
@@ -1,6 +1,11 @@
 const authRouter = require('express').Router();
-const { loginController } = require('../controllers/auth.controller');
-const { validateLoginReq } = require('../middlewares/login.validator');
+const authController = require('../controllers/auth.controller');
+const middlewares = require('../middlewares/login.validator');
+const db = require('../models');
 
-authRouter.post('/login', validateLoginReq, loginController);
+authRouter.post('/login', middlewares.validateLoginReq, authController.loginController);
+authRouter.post('/register', authController.createUserLogin);
+authRouter.get('/', async (req, res) => {
+  res.send(await db.auth.findAll());
+});
 module.exports = authRouter;
