+++ src/routes/user.route.js
@@ -1,13 +1,11 @@
 const express = require('express');
-const { listUsers, postUser, deleteUser } = require('../controllers/user.controller');
+const { listUsers, createUser, deleteUser } = require('../controllers/user.controller');
 const validator = require('../middlewares/request.validator');
-// require schemas
 const schemas = require('../middlewares/schemas.validator');
-
 const router = express.Router();
 
-router.get('/users', listUsers);
-router.post('/users', validator(schemas.userSchema, 'body'), postUser);
-router.delete('/users/:id', deleteUser);
+router.get('/users', validator.reqAuthValidator,listUsers);
+router.post('/users', validator.reqAuthValidator, createUser);
+router.delete('/users/:id', validator.reqAuthValidator,deleteUser);
 
 module.exports = router;
