+++ tests/utils/token.verification.test.js
@@ -1,6 +1,6 @@
 const db = require('../../src/models');
 const LoginError = require('../../src/utils/LoginError');
-const tokenVerification = require('../../src/utils/tokenVerification');
+const tokenVerification = require('../../src/utils/token.verification');
 const jwt = require('jsonwebtoken');
 describe('Check tokenVerification Util', () => {
   it('check if the token is getting verified or not', async () => {
@@ -10,7 +10,6 @@ describe('Check tokenVerification Util', () => {
       exp: 1675972409,
     });
     jest.spyOn(db.users, 'findOne').mockResolvedValue({
-      dataValues: {
         user_id: '9a492c13-85e8-4b26-9339-a5d037664d1a',
         name: 'Promit Revar',
         email: 'promit.revar2211@gmail.com',
@@ -24,7 +23,6 @@ describe('Check tokenVerification Util', () => {
         image: null,
         createdAt: '2023-02-09T15:02:53.658Z',
         updatedAt: '2023-02-09T15:02:53.658Z',
-      },
     });
 
     const result = await tokenVerification.verifyToken('token');
